added RoundedMoney newtype for transporter BPP

+++ app/driver-offer-bpp/src/Core/ACL/OnConfirm.hs
@@ -14,7 +14,7 @@ mkOnConfirmMessage now res = do
   let vehicleVariant = Common.castVariant res.booking.vehicleVariant
   let itemCode = OnConfirm.ItemCode OnConfirm.ONE_WAY_TRIP vehicleVariant Nothing Nothing
       fareParams = booking.fareParams
-      totalFareDecimal = realToFrac booking.estimatedFare
+      totalFareDecimal = fromIntegral booking.estimatedFare
       currency = "INR"
   OnConfirm.OnConfirmMessage
     { order =
@@ -33,7 +33,7 @@ mkOnConfirmMessage now res = do
                       },
                   breakup =
                     mkBreakupList
-                      (OnConfirm.BreakupItemPrice currency . realToFrac)
+                      (OnConfirm.BreakupItemPrice currency . fromIntegral)
                       OnConfirm.BreakupItem
                       fareParams
                 },
